requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
babel==2.17.0
bcrypt==4.0.1
certifi==2025.8.3
cffi==1.17.1
click==8.1.8
cloudinary==1.44.1
cryptography==44.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.116.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.2
packaging==24.2
passlib==1.7.4
pendulum==3.1.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.5
pydantic-partial==0.7.0
pydantic-settings==2.10.1
pydantic_core==2.27.2
Pygments==2.19.2
PyJWT==2.10.1
pypika-tortoise==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
starlette-admin==0.16.0
tortoise-orm==0.25.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.3
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
WTForms==3.1.2
alembic
